requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.7
click==8.3.1
cmake==4.3.0
colorama==0.4.6
Django==5.2.12
dlib-bin==20.0.1
dotenv==0.9.9
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.24.4
opencv-python==4.8.1.78
pillow==12.1.1
python-dotenv==1.2.2
reportlab==4.4.10
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
